Problem: depending one external issue trackers has downsides
In certain projects, issues are not just todo-lists that can
be discarded once in a while. In fact, they contain some vital
records on how decisions are made, giving a unique insight
into the reasoning behind them.

However, most of popular options are not great for that. Over time,
services shut down, SCMs go out of fashion, etc.

Solution: design an SCM-agnostic, never-delete-anything issue
tracker.

This commit includes a very basic prototype to start off. It is
not the most efficient or ergonomic implementation. The code can
also use a lot of improvements. There are bugs for sure. You have
been warned. Good luck!