Problem: no record authorship preserved
SIT itself can't tell who created a record and the only way to retrieve
this information is to use SCM's functionality (such as `git blame`) to
figure that out. That's less than ideal, especially considering that
this information will not surive a loss of an
SCM repository.

Solution: introduce a convention for recording authors' information
and make `sit` tool do this by default