Problem: querying information from records
Currently, this would require scanning the contents of
.sit. While not a terribly bad thing, it'd be completely
outside of sit's tooling and therefore, again, everybody
would have to come up with their own solution.

Solution: add JMESPath filtering and querying to `records`
subcommand.

This is very similar to JMESPath filter and querying in
`issues`.

Also, this makes `Record` serde-serializable.

Currently, handling of binary files isn't best. Should it
base64 instead of just letting people know it is binary?

Or should it sniff the MIME type?