Problem: reducers state being non-map is confusing
Basically, as far as the whole convention of reducers
if concerned, the state is a JSON object. It doesn't
make a lot of sense to allow other types (such as string,
or number) to represent a reduction. While it might have
negligible benefit to special case reduction where the
result is a single type, this affects the code of existing
regular reducers as they have to handle the fact
that incoming state might be any JSON type.

Solution: reduce the scope to JSON objects only
at the type level.