Problem: hashing order is not deterministic
Even though it says it is in the README, it is not.
I simply forgot to implement that part. Mea culpa.

Solution: make hashing order deterministic

The files are sorted by their path as UTF-8
lexicographically now (on Windows, backslashes are
replaced with forward slashes for the purpose of hashing)

While the fix itself was relatively simple,
there was a bigger issue: what should we do
with the existing database of issues? Clearly,
the hashes for some (many?) of these are not
going to match the deterministic ones.

As a solution, I've added `sit rebuild` command
which allows to rebuild a repository to ensure
correct hashes are used. To expand on its usability,
an option has been added to allow running a command
on every record. This will allow us doing things
like fixing invalid records or upgrading conventions
easily.