Problem: Web UI does not allow to edit the title or summary
There has been some reluctance to add this because there's still
no permissions management convention in SIT (who is allowed
to override items like this?)

If some reasonable rules as two who is allowed to update in
absence of permissioning can be formalized, then this should be done.

Solution: forego the reluctance, repurpose "New issue" button
and enable editing issue's summary and details.

Right now this will allow anybody to edit it, as is really the
case with the rest of the system (lack of permissioning). I think
this is fine. Once the permissioning convention is in place
(whatever that system is, as it might vary in different settings),
unallowed edits will be discarded by reducers and we can also let
the reducers produce "who can edit/you can edit" outputs in issue's
view.