Problem: Travis CI fails to push out a CI report
It seems like the fork point for the CI report branch
includes the patch itself

Solution: checkout a `ci` branch before performing CI

./scripts/ci requires reference as a "current branch"
to return to and checkout CI report branch from.

However, Travis CI checks out a detached commit. This
makes the script assume an empty string as the current
branch. When it tries to switch back to it, it's effectively
`git checkout` and so the script remains in the branch
with the merge request.

By checking out `ci` branch we satisfy this condition.