Problem: inability to quickly tell the status of MRs
Some of the merge requests that were not merged in eagerly
(being too experimental for that being the most common cause)
tend to diverge from the master but there's no easy way to
tell this until attempted.

Solution: provide a script (./scripts/mr-report) to list MRs

It will show if they have conflicts and show the patches with
conflicts. This is not a perfect solution yet (ideally it should
be integrated into issue-tracking module and perhaps appear
in the UI as well, if feasible), but this solves the problem
here and now.