Problem: padding in base32 encoding
I've noticed that currently used Base32 encoding has padding enabled.
While this is not harmful per se, the non-alphanumeric character of "="
might be problematic on some file systems and in URLs.

Solution: switch the padding character to '-'

Luckily for us, we've only used 20-byte long Blake2 and SHA-1 hashing
algorithms so far so SIT has never produced padded record names (that
is, unless somebody used SIT with a manually modified config to state
a different length of the Blake2 algorithm output -- which is highly
unlikely at this stage of SIT's adoption).