Problem: sit web doesn't render files inside item/ITEM/RECORD
It can't find the file through directory resolution

Solution: improve `resolve_dir` implementation

Recent change that added mandatory `root` argument to `resolve_dir` has
a bug: when it needs to go through every component of the path
(necessary when trying to find links inside of a longer path) it'll
recurse with the same `root` value as passed.

However, even the very first component is (on UNIX) a root directory
which clearly violates `root` restriction if it is any narrower than
root directory, which will return a `NotFound` error.

In case of sit web issue, `resolve_dir` was called with repository's
path as `root`. Because of the above issue, `NotFound` was returned on
legitimate paths.

The fix is simple: always use root directory `root` restriction
internally and rely on prefix stripping at the end of the process.